% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeupper.R
\name{makeupper}
\alias{makeupper}
\title{Make all column names in uppercase}
\usage{
makeupper(i)
}
\arguments{
\item{i}{dataframe where to make all columns names in uppercase}
}
\value{
dataframe with all columns in uppercase
}
\description{
Make all column names in uppercase
}
\examples{

}
