% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_standardize.R
\name{to_standardize}
\alias{to_standardize}
\title{To standardize a numerical variable}
\usage{
to_standardize(x)
}
\arguments{
\item{x}{variable to standardize}
}
\value{
column with variable standardize
}
\description{
To standardize a numerical variable
}
\examples{

}
